refactor: signer mgmt
