+++ src/tests/btc2x.test.ts
@@ -1,18 +1,23 @@
+import { getTokenByChainAndSymbol } from '@indexcoop/tokenlists'
 import {
   getMainnetTestFactory,
   getMainnetTestFactoryUniswap,
+  getSignerAccount,
+  getTestRpcProvider,
   QuoteTokens,
-  SignerAccount4,
   wei,
 } from './utils'
 
 import type { TestFactory } from './utils'
 
-const { btc2x, eth } = QuoteTokens
-
 describe('BTC2X (mainnet)', () => {
-  const signer = SignerAccount4
+  const chainId = 1
+  const btc2x = getTokenByChainAndSymbol(chainId, 'BTC2X')
+  const { eth } = QuoteTokens
+
   let factory: TestFactory
+  const provider = getTestRpcProvider(chainId)
+  const signer = getSignerAccount(4, provider)
   beforeEach(async () => {
     factory = getMainnetTestFactory(signer)
   })
